feat: created directory to store interfaces and types
