+++ frontend/interfaces/Medication.tsx
@@ -1,0 +1,7 @@
+export interface Medication {
+  id: string;
+  name: string;
+  description: string;
+  dosage: string;
+  category: string;
+}
